added backend security to the order routes

+++ client/store/product.js
@@ -41,6 +41,7 @@ export const GetCartLocally = (array, total) => ({
 })
 export const DeleteItemLocally = () => ({type: DELETE_ITEM_LOCALLY})
 
+///Protected
 export const getCart = () => async dispatch => {
   try {
     const {data} = await axios.get('/api/orders')
@@ -49,7 +50,7 @@ export const getCart = () => async dispatch => {
     console.log(error)
   }
 }
-
+///Protected
 export const updateCart = (product, itemQty) => async dispatch => {
   try {
     await axios.post(`/api/itemizeds`, {product, itemQty})
@@ -95,7 +96,7 @@ export const updateProduct = (id, product) => async dispatch => {
     console.error(err)
   }
 }
-
+//Nopt sure if protected
 export const updateQtyItem = (itemQty, product) => async dispatch => {
   try {
     await axios.put('/api/itemizeds/updateQty', {
@@ -107,19 +108,20 @@ export const updateQtyItem = (itemQty, product) => async dispatch => {
     console.log(error)
   }
 }
-
-export const deleteItem = (productId, orderId) => async dispatch => {
+///Protected
+export const deleteItem = (productId, orderId, userId) => async dispatch => {
   try {
-    await axios.delete(`/api/itemizeds/${productId}/${orderId}`)
+    await axios.delete(`/api/itemizeds/${productId}/${orderId}/${userId}`)
     dispatch(getCart())
   } catch (error) {
     console.log(error)
   }
 }
-
-export const purchaseOrder = user => async dispatch => {
+///Protected
+export const purchaseOrder = (user, userId) => async dispatch => {
   try {
-    const {data} = await axios.put('api/orders', {user})
+    console.log('ORDER', user)
+    const {data} = await axios.put(`api/orders/${userId}`, {user})
     dispatch(Purchase(data))
   } catch (error) {
     console.error(error)
